init suport for org mode
