+++ modules/markup/markdown/markdown.go
@@ -1,0 +1,100 @@
+// Copyright 2014 The Gogs Authors. All rights reserved.
+// Copyright 2014 The Gitea Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package markdown
+
+import (
+	"strings"
+
+	"code.gitea.io/gitea/modules/markup"
+	"code.gitea.io/gitea/modules/setting"
+
+	"github.com/russross/blackfriday"
+)
+
+// IsMarkdownFile reports whether name looks like a Markdown file
+// based on its extension.
+func IsMarkdownFile(name string) bool {
+	return markup.IsMarkupFile(name, MarkupName)
+}
+
+// RenderRaw renders Markdown to HTML without handling special links.
+func RenderRaw(body []byte, urlPrefix string, isWiki bool) []byte {
+	htmlFlags := 0
+	htmlFlags |= blackfriday.HTML_SKIP_STYLE
+	htmlFlags |= blackfriday.HTML_OMIT_CONTENTS
+	renderer := &Renderer{
+		Renderer:  blackfriday.HtmlRenderer(htmlFlags, "", ""),
+		URLPrefix: urlPrefix,
+		IsWiki:    isWiki,
+	}
+
+	// set up the parser
+	extensions := 0
+	extensions |= blackfriday.EXTENSION_NO_INTRA_EMPHASIS
+	extensions |= blackfriday.EXTENSION_TABLES
+	extensions |= blackfriday.EXTENSION_FENCED_CODE
+	extensions |= blackfriday.EXTENSION_STRIKETHROUGH
+	extensions |= blackfriday.EXTENSION_NO_EMPTY_LINE_BEFORE_BLOCK
+
+	if setting.Markdown.EnableHardLineBreak {
+		extensions |= blackfriday.EXTENSION_HARD_LINE_BREAK
+	}
+
+	body = blackfriday.Markdown(body, renderer, extensions)
+	return body
+}
+
+// Render renders Markdown to HTML with all specific handling stuff.
+func Render(rawBytes []byte, urlPrefix string, metas map[string]string) []byte {
+	return render(rawBytes, urlPrefix, metas, false)
+}
+
+// RenderString renders Markdown to HTML with special links and returns string type.
+func RenderString(raw, urlPrefix string, metas map[string]string) string {
+	return string(Render([]byte(raw), urlPrefix, metas))
+}
+
+// RenderWiki renders markdown wiki page to HTML and return HTML string
+func RenderWiki(rawBytes []byte, urlPrefix string, metas map[string]string) string {
+	return string(render(rawBytes, urlPrefix, metas, true))
+}
+
+// render renders Markdown to HTML with all specific handling stuff.
+func render(rawBytes []byte, urlPrefix string, metas map[string]string, isWikiMarkdown bool) []byte {
+	urlPrefix = strings.Replace(urlPrefix, " ", "+", -1)
+	result := RenderRaw(rawBytes, urlPrefix, isWikiMarkdown)
+	result = markup.PostProcess(result, urlPrefix, metas, isWikiMarkdown)
+	result = markup.SanitizeBytes(result)
+	return result
+}
+
+var (
+	// MarkupName describes markup's name
+	MarkupName = "markdown"
+)
+
+func init() {
+	markup.RegisterParser(Parser{})
+}
+
+// Parser implements markup.Parser
+type Parser struct {
+}
+
+// Name implements markup.Parser
+func (Parser) Name() string {
+	return MarkupName
+}
+
+// Extensions implements markup.Parser
+func (Parser) Extensions() []string {
+	return setting.Markdown.FileExtensions
+}
+
+// Render implements markup.Parser
+func (Parser) Render(rawBytes []byte, urlPrefix string, metas map[string]string, isWiki bool) []byte {
+	return RenderRaw(rawBytes, urlPrefix, isWiki)
+}
