+++ modules/markup/orgmode/orgmode_test.go
@@ -1,0 +1,5 @@
+// Copyright 2017 The Gitea Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package markup
